Refactor Insert Commands, Update Room Cleaning Charges, Add View Bills Page
✅ Refactored all insert operations from OpenAnimation into AppDatabase for proper architecture and consistency.
✅ Updated Room Cleaning Service to charge RM15 starting from the second request onwards (first one still free because we’re sweet like that 💅).
✅ Added a brand new View Bills Page to display outstanding service charges and allow users to track their spending!

--- Yuki
