+++ app/src/main/res/layout/activity_view_bills.xml
@@ -15,19 +15,30 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent" />
 
-    <!-- Content Area -->
+    <!-- Default Message -->
     <TextView
-        android:id="@+id/serviceContent"
+        android:id="@+id/noOutstandingBillsText"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="View Bills"
+        android:text="N O   O U T S T A N D I N G   B I L L S"
         android:textColor="@android:color/white"
         android:textSize="18sp"
+        android:visibility="gone"
         app:layout_constraintTop_toBottomOf="@id/header_container"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintBottom_toTopOf="@id/footer_container" />
+
+    <!-- RecyclerView for Service Requests -->
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/billsRecyclerView"
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        android:padding="16dp"
+        app:layout_constraintTop_toBottomOf="@id/header_container"
         app:layout_constraintBottom_toTopOf="@id/footer_container"
-        android:layout_margin="16dp" />
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent" />
 
     <!-- Footer -->
     <com.example.testversion.FooterBar
